requirements.txt
blinker==1.5
certifi==2021.10.8
click==8.1.7
colorama==0.4.6
dnspython==2.2.1
flask==2.2.2
importlib-metadata==5.0.0
itsdangerous==2.1.2
jinja2==3.1.2
MarkupSafe==2.1.1
pymongo==4.3.2
python-dotenv==0.21.0
werkzeug==2.2.2
zipp==3.10.0

